+++ include/dimensional/si/kelvin.hpp
@@ -9,6 +9,10 @@ struct thermodynamic_temperature {
 };
 
 using kelvin_t = make_unit_t<thermodynamic_temperature>;
+using millikelvin_t = scaled_unit_t<kelvin_t, std::milli>;
+using microkelvin_t = scaled_unit_t<kelvin_t, std::milli>;
+using kilokelvin_t = scaled_unit_t<kelvin_t, std::kilo>;
+using megakelvin_t = scaled_unit_t<kelvin_t, std::mega>;
 
 inline constexpr kelvin_t kelvins{};